BE/requirements.txt
fastapi
pydantic
requests
uvicorn
python-decouple
starlette
python-dotenv~=1.0.0
python-multipart
passlib
PyJWT
pandas
bcrypt
debugpy
email-validator
sqlalchemy
psycopg2
cachetools==5.5.2
google-api-python-client==2.166.0
logging==0.4.9.6
typing-extensions==4.13.0
ytmusicapi==1.10.2
elevenlabs==1.55.0
google-generativeai==0.8.4